Introduce local development server
This commit provides users the quick way to run samples in a
way already mentioned in README.md:

- start a local server
- a web page should be opened ready to use samples starting
on the index page

This is achieved with help of NPM installed `browser-sync`
tool:
https://www.browsersync.io/

The tool is installed as local dependency using NPM and just
starts when using default NPM start command using the same
UX scenario as mentioned in changed README content

The `package.json` file is being added to repository content
as poart of this commit.

Thanks!